@@ -16,116 +30,45 @@ printf "Host: %s\nOS: %s %s %s\nJVM: %s %s %s %s\nGroovy: %s\nGradle: %s\n" +
         System.getProperty("java.vm.version"),
         System.getProperty("java.vm.vendor"),
         System.getProperty("java.vm.name"),
-        GroovySystem.getVersion(),
-        gradle.gradleVersion
+        gradle.gradleVersion, GroovySystem.getVersion(), JavaVersion.current()
 
 apply plugin: 'java'
 apply plugin: 'maven'
 apply plugin: 'signing'
-apply plugin: 'findbugs'
-apply plugin: 'pmd'
-apply plugin: 'checkstyle'
-apply plugin: "jacoco"
+apply plugin: 'org.xbib.gradle.plugin.elasticsearch.plugin'
 apply plugin: 'org.xbib.gradle.plugin.asciidoctor'
 
-repositories {
-    mavenCentral()
+esplugin {
+    description 'Decompound plugin for Elasticsearch'
+    classname 'org.xbib.elasticsearch.plugin.analysis.decompound.AnalysisDecompoundPlugin'
+    name 'elasticsearch-analysis-decompound'
+    url 'https://github.com/jprante/elasticsearch-analysis-decompound'
 }
 
 configurations {
     wagon
-    distJars {
-        extendsFrom runtime
-        exclude group: 'org.elasticsearch'
-        exclude module: 'lucene-core'
-        exclude module: 'lucene-analyzers-common'
-        exclude module: 'jna'
-        exclude module: 'jackson-core'
-        exclude module: 'jackson-dataformat-smile'
-        exclude module: 'jackson-dataformat-yaml'
-        exclude module: 'log4j-api'
-    }
 }
 
-apply from: 'gradle/ext.gradle'
-apply from: 'gradle/publish.gradle'
-apply from: 'gradle/sonarqube.gradle'
-
 dependencies {
-    compile "org.elasticsearch:elasticsearch:${project.property('elasticsearch.version')}"
-    compile "org.apache.logging.log4j:log4j-api:${project.property('log4j.version')}"
-    testCompile "junit:junit:${project.property('junit.version')}"
     testCompile "org.apache.logging.log4j:log4j-core:${project.property('log4j.version')}"
     testCompile "org.elasticsearch.plugin:transport-netty4-client:${project.property('elasticsearch.version')}"
-    distJars "${project.group}:${project.name}:${project.version}"
     wagon "org.apache.maven.wagon:wagon-ssh:${project.property('wagon.version')}"
 }
 
-sourceCompatibility = JavaVersion.VERSION_1_8
+sourceCompatibility = JavaVersion.VERSION_1_9
 targetCompatibility = JavaVersion.VERSION_1_8
 
-[compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
-tasks.withType(JavaCompile) {
-    options.compilerArgs << "-Xlint:all" << "-profile" << "compact2"
-}
+esTest.dependsOn jar
 
-test {
-    systemProperties['path.home'] = System.getProperty("user.dir")
-    testLogging {
-        showStandardStreams = false
-        exceptionFormat = 'full'
-    }
-}
-
-task makePluginDescriptor(type: Copy) {
-    from 'src/main/templates'
-    into 'build/tmp/plugin'
-    expand([
-            'descriptor': [
-                    'name': pluginName,
-                    'classname': pluginClassname,
-                    'description': pluginDescription,
-                    'version': project.property('version'),
-                    'javaVersion': project.property('targetCompatibility'),
-                    'elasticsearchVersion' : project.property('elasticsearch.version')
-            ]
-    ])
-}
-
-task buildPluginZip(type: Zip, dependsOn: [':jar', ':makePluginDescriptor']) {
-    from configurations.distJars
-    from 'build/tmp/plugin'
-    into 'elasticsearch'
-    classifier = 'plugin'
-}
-
-task unpackPlugin(type: Copy, dependsOn: [':buildPluginZip']) {
-    delete "plugins"
-    from configurations.distJars
-    from 'build/tmp/plugin'
-    into "plugins/${pluginName}"
-}
+test.enabled = false
+randomizedTest.enabled = false
 
 clean {
-    delete "plugins"
-    delete "data"
-    delete "logs"
-}
-
-task javadocJar(type: Jar, dependsOn: classes) {
-    from javadoc
-    into 'build/tmp'
-    classifier 'javadoc'
-}
-
-task sourcesJar(type: Jar, dependsOn: classes) {
-    from sourceSets.main.allSource
-    into 'build/tmp'
-    classifier 'sources'
+    delete fileTree('.') { include '.local*.log' }
 }
 
 artifacts {
-    archives sourcesJar, javadocJar, buildPluginZip
+    archives sourcesJar, javadocJar
 }
 
 if (project.hasProperty('signing.keyId')) {
